feat(event): using INT instead of SMALLINT for count

+++ pkg/persistence/migrator.go
@@ -95,11 +95,11 @@ CREATE TABLE IF NOT EXISTS failed_key_claim_attempts (
 		statements: []string{
 			`
 CREATE TABLE IF NOT EXISTS events (
-	source		VARCHAR(32)  NOT NULL,
-	identifier 	VARCHAR(255) NOT NULL,
-	device_type	VARCHAR(32) NOT NULL,
-	date 				DATE NOT NULL,
-	count 			SMALLINT UNSIGNED NOT NULL DEFAULT 0,
+	source          VARCHAR(32)     NOT NULL,
+	identifier      VARCHAR(255)    NOT NULL,
+	device_type     VARCHAR(32)     NOT NULL,
+	date            DATE            NOT NULL,
+	count           INT             UNSIGNED NOT NULL DEFAULT 0,
 	INDEX (source),
 	INDEX (identifier),
 	INDEX (device_type),
